Refactor backend directory
