+++ backend/main.ts
@@ -1,0 +1,40 @@
+// Modules to control application life and create native browser window
+const { app, protocol, BrowserWindow } = require('electron')
+const path = require('path')
+
+require('./server')
+
+const createWindow = () => {
+  // Create the browser window.
+  const mainWindow = new BrowserWindow({
+    icon: 'public/favicon.ico',
+    width: 1024,
+    height: 920,
+    webPreferences: {
+      preload: path.join(__dirname, 'preload.ts')
+    }
+  })
+
+  // and load the index.html of the app.
+  mainWindow.loadFile('dist/index.html')
+}
+
+// This method will be called when Electron has finished
+// initialization and is ready to create browser windows.
+// Some APIs can only be used after this event occurs.
+app.whenReady().then(() => {
+  createWindow()
+
+  app.on('activate', () => {
+    // On macOS it's common to re-create a window in the app when the
+    // dock icon is clicked and there are no other windows open.
+    if (BrowserWindow.getAllWindows().length === 0) createWindow()
+  })
+})
+
+// Quit when all windows are closed, except on macOS. There, it's common
+// for applications and their menu bar to stay active until the user quits
+// explicitly with Cmd + Q.
+app.on('window-all-closed', () => {
+  if (process.platform !== 'darwin') app.quit()
+})
